requirements.txt
aiofiles
aiohttp
aiosignal
asgiref
async-timeout
attrs
better-ipc
blinker
certifi
charset-normalizer
click
colorama
discord.py
Django
exceptiongroup
Flask
frozenlist
h11
h2
hpack
Hypercorn
hyperframe
idna
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mdurl
multidict
numpy
priority
Pygments
requests
rich
setuptools
shellingham
sqlparse
taskgroup
thread
tomli
typer
typing_extensions
tzdata
urllib3
websockets
Werkzeug
wsproto
yarl
